requirements.txt
Django==4.2.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
whitenoise==6.4.0